requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary
pydantic==2.5.0
coverage==7.3.2
python-multipart==0.0.6 
pandas
requests
httpx
numpy 